Fix Warning Message

+++ src/components/admin/category/categoryLists.js
@@ -41,9 +41,8 @@ class categoryLists extends Component {
     this.setState( { loading: true}, () => {
       commonService.getAPIWithAccessToken('category')
         .then( res => {
-          console.log(res);
            
-          if ( undefined == res.data.data || !res.data.status ) {
+          if ( undefined === res.data.data || !res.data.status ) {
             this.setState( {  loading: false } );
             toast.error(res.data.message);    
             return;
@@ -53,7 +52,7 @@ class categoryLists extends Component {
          
         } )
         .catch( err => {         
-          if(err.response.status === 401 && err.response.status !== undefined) {
+          if(err.response !== undefined && err.response.status === 401) {
             localStorage.clear();
             this.props.history.push('/login');
           }
@@ -100,7 +99,7 @@ class categoryLists extends Component {
           } )
           .catch( err => {       
               
-            if(err.response.status == 401) {
+            if(err.response !== undefined && err.response.status === 401) {
               localStorage.clear();
               this.props.history.push('/login');
             }
@@ -130,7 +129,7 @@ class categoryLists extends Component {
           } )
           .catch( err => {       
               
-            if(err.response.status == 401) {
+            if(err.response !== undefined && err.response.status === 401) {
               localStorage.clear();
               this.props.history.push('/login');
             }
@@ -224,7 +223,7 @@ class categoryLists extends Component {
         } )
         .catch( err => {       
             
-          if(err.response.status == 401) {
+          if(err.response !== undefined && err.response.status === 401) {
             localStorage.clear();
             this.props.history.push('/login');
           }
